eth provider: split provider in multiple traits
