+++ src/providers/alchemy_provider.rs
@@ -3,7 +3,7 @@ use crate::{
     providers::eth_provider::{
         contracts::erc20::EthereumErc20,
         error::EthApiError,
-        provider::{EthProviderResult, EthereumProvider},
+        provider::{EthProviderResult, EthereumProvider1},
     },
 };
 use async_trait::async_trait;
@@ -34,18 +34,18 @@ pub trait AlchemyProvider {
 }
 
 #[derive(Debug, Clone)]
-pub struct AlchemyDataProvider<P: EthereumProvider> {
+pub struct AlchemyDataProvider<P: EthereumProvider1> {
     eth_provider: P,
 }
 
-impl<P: EthereumProvider> AlchemyDataProvider<P> {
+impl<P: EthereumProvider1> AlchemyDataProvider<P> {
     pub const fn new(eth_provider: P) -> Self {
         Self { eth_provider }
     }
 }
 
 #[async_trait]
-impl<P: EthereumProvider + Send + Sync + 'static> AlchemyProvider for AlchemyDataProvider<P> {
+impl<P: EthereumProvider1 + Send + Sync + 'static> AlchemyProvider for AlchemyDataProvider<P> {
     async fn token_balances(
         &self,
         address: Address,
