+++ src/providers/eth_provider/mod.rs
@@ -1,7 +1,15 @@
+pub mod blocks;
+pub mod chain;
 pub mod constant;
 pub mod contracts;
 pub mod database;
 pub mod error;
+pub mod gas;
+pub mod logs;
 pub mod provider;
+pub mod receipts;
 pub mod starknet;
+pub mod state;
+pub mod transactions;
+pub mod tx_pool;
 pub mod utils;
